requirements.txt
streamlit==1.32.0
google-generativeai>=0.3.1
python-dotenv==1.0.0
PyMuPDF==1.23.5
Pillow==10.0.0